requirement.txt
asgiref==1.1.2
attrs==17.3.0
autobahn==17.10.1
Automat==0.6.0
channels==1.1.8
channels-api==0.4.0
constantly==15.1.0
daphne==1.3.0
Django==1.11
django-filter==1.1.0
djangorestframework==3.7.3
hyperlink==17.3.1
incremental==17.5.0
Markdown==2.6.10
pkg-resources==0.0.0
pytz==2017.3
six==1.11.0
Twisted==17.9.0
txaio==2.8.2
zope.interface==4.4.3
